requirements.txt
pytest==5.3.5
requests==2.22.0
allure-pytest==2.8.10